+++ packages/magnitude-core/src/web/shadowDOMInputAdapter.d.ts
